% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion-test.R
\name{check_assumptions}
\alias{check_assumptions}
\title{Check the assumptions of the one sample proportion test}
\usage{
check_assumptions(data_vec)
}
\arguments{
\item{data_vec}{A vector of TRUE/FALSE or 0/1 data representing the sample}
}
\value{
No return value, but prints information about the conditions
}
\description{
Check the assumptions of the one sample proportion test
}
