% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T-Test.R
\name{check_outliers}
\alias{check_outliers}
\title{Check for outliers in the data}
\usage{
check_outliers(data)
}
\arguments{
\item{data}{A numeric vector of data.}
}
\value{
A dataframe containing the indices and values of the outliers.
}
\description{
This function identifies outliers in a numeric vector using the interquartile
range (IQR) method.
}
