% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T-Test.R
\name{display_qqplot}
\alias{display_qqplot}
\title{Display a QQ plot for checking the normality assumption}
\usage{
display_qqplot(data, x)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{x}{The name of the column to be checked (must be numeric).}
}
\value{
A QQ plot.
}
\description{
This function generates a QQ plot to visually assess the normality of the data
in the specified column of the dataframe.
}
