% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion-test.R
\name{get_conf_int}
\alias{get_conf_int}
\title{Calculate a confidence interval for a specified level of confidence}
\usage{
get_conf_int(data_vec, conf_level)
}
\arguments{
\item{data_vec}{A vector of TRUE/FALSE or 0/1 data representing the sample}

\item{conf_level}{A number between 0 and 1 specifying the confidence level}
}
\value{
A dataframe with confidence level, lower bound, and upper bound
}
\description{
Calculate a confidence interval for a specified level of confidence
}
