% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion-test.R
\name{get_summary_statistics}
\alias{get_summary_statistics}
\title{Calculates summary statistics for a sample of binary categorical data}
\usage{
get_summary_statistics(data_vec, hypo_p)
}
\arguments{
\item{data_vec}{A vector of TRUE/FALSE or 0/1 data representing the sample}

\item{hypo_p}{Optionally, specify a hypothesized value for the proportion}
}
\value{
A dataframe with summary statistics
}
\description{
Calculates summary statistics for a sample of binary categorical data
}
