% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T-Test.R
\name{get_t_test_result}
\alias{get_t_test_result}
\title{Calculate the result of a one-sample t-test for a population mean}
\usage{
get_t_test_result(
  sample_mean,
  sample_sd,
  n,
  hypo_mean,
  hypo_direction = "two.sided"
)
}
\arguments{
\item{sample_mean}{The sample mean.}

\item{sample_sd}{The sample standard deviation.}

\item{n}{The sample size.}

\item{hypo_mean}{The hypothesized value of the mean under the null hypothesis.}

\item{hypo_direction}{The direction of the alternative hypothesis ("<", ">", "!=").}
}
\value{
A dataframe containing the test statistic (t-value) and p-value.
}
\description{
This function computes the result of a one-sample t-test for a population mean
using the sample mean, sample standard deviation, sample size, hypothesized
mean, and direction of the alternative hypothesis.
}
