% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T-Test.R, R/proportion-test.R
\name{plot_sampling_distribution}
\alias{plot_sampling_distribution}
\title{Plot the sampling distribution for a one-sample t-test}
\usage{
plot_sampling_distribution(p_hat, SE_p_hat, hypo_p, hypo_direction)

plot_sampling_distribution(p_hat, SE_p_hat, hypo_p, hypo_direction)
}
\arguments{
\item{p_hat}{The sample proportion}

\item{SE_p_hat}{The standard error of the sample proportion}

\item{hypo_p}{The hypothesized value of the proportion}

\item{hypo_direction}{The direction of the alternative hypothesis}

\item{sample_mean}{The sample mean.}

\item{sample_sd}{The sample standard deviation.}

\item{n}{The sample size.}

\item{hypo_mean}{The hypothesized population mean.}
}
\value{
A plot of the sampling distribution.

A plot of the sampling distribution
}
\description{
This function creates a plot of the sampling distribution for a one-sample
t-test based on the sample mean, sample standard deviation, sample size,
hypothesized mean, and direction of the alternative hypothesis.
}
