% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T-Test.R
\name{plot_sampling_distribution_t}
\alias{plot_sampling_distribution_t}
\title{Plot the sampling distribution for a one-sample t-test}
\usage{
plot_sampling_distribution_t(
  sample_mean,
  sample_sd,
  n,
  hypo_mean,
  hypo_direction
)
}
\arguments{
\item{sample_mean}{The sample mean.}

\item{sample_sd}{The sample standard deviation.}

\item{n}{The sample size.}

\item{hypo_mean}{The hypothesized population mean.}

\item{hypo_direction}{The direction of the alternative hypothesis ("greater", "less", "two.sided").}
}
\value{
A plot of the sampling distribution.
}
\description{
This function creates a plot of the sampling distribution for a one-sample
t-test based on the sample mean, sample standard deviation, sample size,
hypothesized mean, and direction of the alternative hypothesis.
}
