% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{slr}
\alias{slr}
\title{This function performs simple linear regression.}
\usage{
slr(data, x, y, intercept = TRUE)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{x}{Predictor variable}

\item{y}{Response variable}

\item{intercept}{True if the intercept should be used, False otherwise}
}
\value{
Common summary results and assumption checks for simple linear regression.
}
\description{
This function performs simple linear regression.
}
