% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T-Test.R
\name{t_test}
\alias{t_test}
\title{Announces the results of a one-sample t-test for a given dataset and column.}
\usage{
t_test(data, x, hypo_mean = 0, alternative = "two.sided", conf_level = 0.95)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{x}{A string representing the name of the column to be tested.}

\item{hypo_mean}{The hypothesized population mean under the null hypothesis (default is 0).}

\item{alternative}{The direction of the alternative hypothesis ("two.sided", "greater", "less"). Default is "two.sided".}

\item{conf_level}{The confidence level for the confidence interval (default is 0.95).}
}
\value{
A printout of the t-test result,
confidence interval, any outliers, and relevant plots.
}
\description{
This function performs a one-sample t-test to determine if the mean of the
specified column in the dataframe is equal to a hypothesized value. It also
checks the normality of the data, identifies outliers, and plots the sampling
distribution.
}
