% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{test_assumptions}
\alias{test_assumptions}
\title{Checks assumptions for simple linear regression}
\usage{
test_assumptions(model, model.diag.metrics, intercept)
}
\arguments{
\item{model}{lm object of slr model}

\item{model.diag.metrics}{common diagnostic metrics for lm model}

\item{intercept}{True if the intercept should be used, False otherwise}
}
\value{
Plots of assumptions for provided lm model
}
\description{
Checks assumptions for simple linear regression
}
