% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion-test.R
\name{valid_data_vec}
\alias{valid_data_vec}
\title{Checks whether the specified data vector is boolean (TRUE/FALSE) or binary numeric (0/1)}
\usage{
valid_data_vec(data_vec)
}
\arguments{
\item{data_vec}{A vector of data representing the sample}
}
\value{
A logical value indicating if the specified data vector is of the correct data type
}
\description{
Checks whether the specified data vector is boolean (TRUE/FALSE) or binary numeric (0/1)
}
