% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion-test.R
\name{valid_direction}
\alias{valid_direction}
\title{Checks whether the specified hypothesized direction follows the correct form}
\usage{
valid_direction(hypo_direction)
}
\arguments{
\item{hypo_direction}{A string representing the direction of the alternative hypothesis}
}
\value{
A logical value indicating if the specified value matches the allowed options
}
\description{
Checks whether the specified hypothesized direction follows the correct form
}
