% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion-test.R
\name{z_test}
\alias{z_test}
\title{Calculate the result of a one-sample z test for a population proportion}
\usage{
z_test(data_vec, hypo_p = -1, hypo_direction = "", conf_level = 0.95)
}
\arguments{
\item{data_vec}{A vector of TRUE/FALSE or 0/1 data representing the sample}

\item{hypo_p}{The hypothesized value of the proportion}

\item{hypo_direction}{The direction of the alternative hypothesis}

\item{conf_level}{A number between 0 and 1 specifying the confidence level}
}
\value{
Several test results, confidence interval, plot of sampling distribution
}
\description{
Calculate the result of a one-sample z test for a population proportion
}
