% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{create_scatter_plot}
\alias{create_scatter_plot}
\title{Creates scatter plot for simple linear regression}
\usage{
create_scatter_plot(df, x, y, model, model.diag.metrics, intercept)
}
\arguments{
\item{df}{A dataframe containing the data.}

\item{x}{Predictor variable}

\item{y}{Response variable}

\item{model}{lm object of slr model}

\item{model.diag.metrics}{common diagnostic metrics for lm model}

\item{intercept}{True if the intercept should be used, False otherwise}
}
\value{
Scatter plot for provided lm model
}
\description{
Creates scatter plot for simple linear regression
}
