% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion-test.R
\name{get_test_result}
\alias{get_test_result}
\title{Calculate the result of a one-sample z test for a population proportion}
\usage{
get_test_result(p_hat, SE_p_hat, hypo_p, hypo_direction)
}
\arguments{
\item{p_hat}{The sample proportion}

\item{SE_p_hat}{The standard error of the sample proportion}

\item{hypo_p}{The hypothesized value of the proportion}

\item{hypo_direction}{The direction of the alternative hypothesis}
}
\value{
A dataframe containing the test statistic and p-value
}
\description{
Calculate the result of a one-sample z test for a population proportion
}
