% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{perform_slr}
\alias{perform_slr}
\title{Performs simple linear regression}
\usage{
perform_slr(df, x, y, intercept = TRUE)
}
\arguments{
\item{df}{Data frame containing predictor and response variables}

\item{x}{Predictor variable}

\item{y}{Response variable}

\item{intercept}{True if the intercept should be used, False otherwise}
}
\value{
Common summary results from simple linear regression
}
\description{
Performs simple linear regression
}
