% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion-test.R
\name{plot_sampling_distribution}
\alias{plot_sampling_distribution}
\title{Create a plot of the sampling distribution of the sample proportion}
\usage{
plot_sampling_distribution(p_hat, SE_p_hat, hypo_p, hypo_direction)
}
\arguments{
\item{p_hat}{The sample proportion}

\item{SE_p_hat}{The standard error of the sample proportion}

\item{hypo_p}{The hypothesized value of the proportion}

\item{hypo_direction}{The direction of the alternative hypothesis}
}
\value{
A plot of the sampling distribution
}
\description{
Create a plot of the sampling distribution of the sample proportion
}
