% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{test_assumptions}
\alias{test_assumptions}
\title{This function checks assumptions for simple linear regression.}
\usage{
test_assumptions(model, intercept = TRUE)
}
\arguments{
\item{model}{lm object of slr model}

\item{intercept}{True if the intercept should be used, False otherwise}

\item{model.diag.metrics}{Common diagnostic metrics for lm model from broom::agument}
}
\value{
Plots of assumptions for provided lm model and outputs observations to look into for violated assumptions.
}
\description{
This function checks assumptions for simple linear regression.
}
